+++ apps/src/sketchlab/reactFlow/ReactFlowSketchLabView.tsx
@@ -2,21 +2,13 @@ import {useTheme} from '@code-dot-org/component-library/common/contexts';
 import FontAwesomeV6Icon from '@code-dot-org/component-library/fontAwesomeV6Icon';
 import {Button as MuiButton} from '@mui/material';
 import {
-  addEdge,
-  Background,
-  Controls,
-  ReactFlow,
   ReactFlowProvider,
-  useEdgesState,
-  useNodesState,
-  useReactFlow,
   type Edge,
   type Node,
-  type OnConnect,
   type Viewport,
 } from '@xyflow/react';
 import '@xyflow/react/dist/style.css';
-import React, {useCallback, useEffect, useRef, useState} from 'react';
+import React, {useCallback, useEffect, useState} from 'react';
 
 import useThemeSetting from '@cdo/apps/lab2/hooks/useThemeSetting';
 import {useVerticalLayout} from '@cdo/apps/lab2/hooks/useVerticalLayout';
@@ -32,21 +24,8 @@ import {useSources} from '@cdo/apps/lab2/views/SourcesContainer';
 import {commonI18n} from '@cdo/apps/types/locale';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 
-import Toolbar from './components/Toolbar';
-import {
-  DEFAULT_NODE_HEIGHT,
-  DEFAULT_NODE_WIDTH,
-  SAVE_DEBOUNCE_MS,
-} from './constants';
-import ImageNode from './nodes/ImageNode';
-import ShapeNode from './nodes/ShapeNode';
-import TextNode from './nodes/TextNode';
-import {
-  ImageNodeData,
-  ReactFlowSketchLabSources,
-  ShapeNodeData,
-  TextNodeData,
-} from './types';
+import ReactFlowCanvas from './components/ReactFlowCanvas';
+import {ReactFlowSketchLabSources} from './types';
 
 import styles from './react-flow-sketch-lab-view.module.scss';
 
@@ -54,144 +33,10 @@ export const REACT_FLOW_DEFAULT_SOURCES: ReactFlowSketchLabSources = {
   source: {nodes: [], edges: []},
 };
 
-const NODE_TYPES = {
-  shape: ShapeNode,
-  image: ImageNode,
-  text: TextNode,
-};
-
 const MIN_INFO_PANEL_WIDTH = 250;
 const INITIAL_INFO_PANEL_WIDTH = 290;
 const MIN_WORKSPACE_WIDTH = 400;
 const INITIAL_WORKSPACE_WIDTH = 800;
-
-// Offset added per new node so they don't stack exactly on top of each other.
-const NEW_NODE_STAGGER_PX = 20;
-
-// ---- Inner canvas component, keyed by mountKey so it fully remounts ----
-// when sources change externally (page load, version history, start over).
-// This follows the same pattern as ExcalidrawSketchLabView where the
-// Excalidraw component is keyed by excalidrawMountKey.
-
-interface CanvasProps {
-  updateSources: ReturnType<
-    typeof useSources<ReactFlowSketchLabSources>
-  >['updateSources'];
-  initialNodes: Node[];
-  initialEdges: Edge[];
-  initialViewport: Viewport | undefined;
-  colorMode: 'light' | 'dark';
-}
-
-function ReactFlowCanvas({
-  updateSources,
-  initialNodes,
-  initialEdges,
-  initialViewport,
-  colorMode,
-}: CanvasProps) {
-  const [nodes, setNodes, onNodesChange] = useNodesState(initialNodes);
-  const [edges, setEdges, onEdgesChange] = useEdgesState(initialEdges);
-  const [viewport, setViewport] = useState<Viewport | undefined>(
-    initialViewport
-  );
-
-  const {screenToFlowPosition} = useReactFlow();
-  const addedNodeCountRef = useRef(0);
-
-  // Debounced save: sync ReactFlow state back to project sources.
-  const saveTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  useEffect(() => {
-    if (saveTimerRef.current) {
-      clearTimeout(saveTimerRef.current);
-    }
-    saveTimerRef.current = setTimeout(() => {
-      updateSources(prev => ({
-        ...prev,
-        source: {nodes, edges, viewport},
-      }));
-    }, SAVE_DEBOUNCE_MS);
-
-    return () => {
-      if (saveTimerRef.current) {
-        clearTimeout(saveTimerRef.current);
-      }
-    };
-  }, [nodes, edges, viewport, updateSources]);
-
-  const onConnect: OnConnect = useCallback(
-    connection => setEdges(eds => addEdge(connection, eds)),
-    [setEdges]
-  );
-
-  const handleMoveEnd = useCallback((_event: unknown, vp: Viewport) => {
-    setViewport(vp);
-  }, []);
-
-  const handleAddNode = useCallback(
-    (
-      type: 'shape' | 'image' | 'text',
-      data: ShapeNodeData | ImageNodeData | TextNodeData
-    ) => {
-      const stagger = addedNodeCountRef.current * NEW_NODE_STAGGER_PX;
-      addedNodeCountRef.current += 1;
-
-      const position = screenToFlowPosition({
-        x: window.innerWidth / 2 - DEFAULT_NODE_WIDTH / 2 + stagger,
-        y: window.innerHeight / 2 - DEFAULT_NODE_HEIGHT / 2 + stagger,
-      });
-
-      const newNodeId = crypto.randomUUID();
-      const newNode: Node = {
-        id: newNodeId,
-        type,
-        position,
-        data: data as unknown as Node['data'],
-        style: {
-          width: DEFAULT_NODE_WIDTH,
-          height: DEFAULT_NODE_HEIGHT,
-        },
-      };
-
-      setNodes(nds => [...nds, newNode]);
-
-      // Move focus to the new node after React Flow renders it.
-      (document.activeElement as HTMLElement)?.blur();
-      setTimeout(() => {
-        const nodeEl = document.querySelector<HTMLElement>(
-          `.react-flow__node[data-id="${newNodeId}"]`
-        );
-        nodeEl?.focus();
-      }, 100);
-    },
-    [screenToFlowPosition, setNodes]
-  );
-
-  return (
-    <div className={styles.canvasContainer}>
-      <Toolbar onAddNode={handleAddNode} />
-      <ReactFlow
-        nodes={nodes}
-        edges={edges}
-        onNodesChange={onNodesChange}
-        onEdgesChange={onEdgesChange}
-        onConnect={onConnect}
-        nodeTypes={NODE_TYPES}
-        onMoveEnd={handleMoveEnd}
-        defaultViewport={initialViewport}
-        fitView={!initialViewport}
-        colorMode={colorMode}
-        deleteKeyCode="Delete"
-        proOptions={{hideAttribution: true}}
-      >
-        <Background />
-        <Controls />
-      </ReactFlow>
-    </div>
-  );
-}
-
-// ---- Outer component: layout, sources, reinitialization ----
 
 function ReactFlowSketchLabViewInner({
   levelProperties,
